requirements.txt
discord~=1.7.3
requests~=2.27.1
python-dotenv~=0.20.0